+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Added option BUILD_BENCHMARKS for building microbenchmarks
 * Extend Python API of UniformTSDFVolume to allow setting the origin
 * Corrected documentation of PointCloud.h
+* Added ISS Keypoint Detector
 * Added an RPC interface for external visualizers running in a separate process
 
 ## 0.9.0